managed-entitlements/src/requirements.txt
boto3
cryptography
PyJWT
